telegram_bot/bot.py
from telegram import Update
from telegram.ext import Updater, CommandHandler, CallbackContext
import logging

# إعداد التسجيل
logging.basicConfig(format='%(asctime)s - %(name)s - %(levelname)s - %(message)s', level=logging.INFO)

# رمز التوكن الخاص بالروبوت
TELEGRAM_TOKEN = 'YOUR_TELEGRAM_BOT_TOKEN'

# وظيفة بدء تشغيل البوت
def start(update: Update, context: CallbackContext):
    update.message.reply_text('مرحبا! أنا هنا لمساعدتك.')

# وظيفة إيقاف تشغيل البوت
def stop(update: Update, context: CallbackContext):
    update.message.reply_text('إيقاف البوت...')
    # هنا يمكنك إضافة الكود لإيقاف البوت أو إيقاف التداول

# وظيفة إرسال رسالة
def send_message(update: Update, context: CallbackContext):
    message = ' '.join(context.args)
    update.message.reply_text(f'رسالة: {message}')

# وظيفة لتلقي الأوامر
def main():
    updater = Updater(TELEGRAM_TOKEN, use_context=True)
    dp = updater.dispatcher

    # إضافة المعالجات
    dp.add_handler(CommandHandler("start", start))
    dp.add_handler(CommandHandler("stop", stop))
    dp.add_handler(CommandHandler("send", send_message))

    # بدء تشغيل البوت
    updater.start_polling()
    updater.idle()

if __name__ == '__main__':
    main()
