telegram_bot/requirements.txt
python-telegram-bot==13.7
python-telegram-bot
ccxt